requirements.txt
# Core dependencies
aiogram>=3.4.1,<4.0.0
httpx>=0.27.0
pydantic>=2.7.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
structlog>=24.1.0
humanfriendly>=10.0
PyYAML>=6.0
psutil>=5.9.0

# Database (PostgreSQL support)
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Retry and resilience
tenacity>=8.2.0

# AI and enrichment (OpenAI removed)

# Document generation
python-docx>=1.1.0

# FastAPI for Robokassa callbacks
fastapi>=0.104.0
uvicorn>=0.24.0

# Web automation for Gamma PDF extraction
selenium>=4.15.0
webdriver-manager>=4.0.0

# XML processing for SOAP
lxml>=4.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
responses>=0.23.0

# Development
ruff>=0.5.0
black>=24.4.0
mypy>=1.0.0